echo command works, read extended description
enviromentalVars function should really be named something else, 
this function grabs what the user types and works with that. 
so far I have the echo commands working, and it also shows error message when the user types in an invalid command. 
this function needs more work, like adding when the user inputs the command "cd" and other built in commands. 
essentially this function is where we will input everything else needed since its the function that is manipulating what the user types in.